requirements.txt
Flask==3.1.2
mysql-connector-python==9.5.0
pandas==2.3.3
numpy==2.3.5
joblib==1.5.2
python-dotenv==1.2.1

scikit-learn==1.7.2
gunicorn
